brax/learning_brax/cartpole/requirements.txt
absl-py==1.4.0
brax==0.9.0
cached-property==1.5.2
chex==0.1.7
click==8.1.3
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
dm-env==1.6
dm-tree==0.1.8
etils==1.2.0
Flask==2.2.3
Flask-Cors==3.0.10
flax==0.6.8
fonttools==4.39.3
glfw==2.5.9
grpcio==1.53.0
gym==0.26.2
gym-notices==0.0.8
importlib-resources==5.12.0
itsdangerous==2.1.2
jax==0.4.8
jaxlib==0.4.7+cuda11.cudnn86
jaxopt==0.6
Jinja2==3.1.2
kiwisolver==1.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
ml-dtypes==0.0.4
msgpack==1.0.5
mujoco==2.3.3
nest-asyncio==1.5.6
numpy==1.24.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-nvcc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.8.1.3
nvidia-cufft-cu11==10.9.0.58
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
opt-einsum==3.3.0
optax==0.1.4
orbax==0.1.7
packaging==23.0
Pillow==9.5.0
protobuf==3.20.3
Pygments==2.15.0
PyOpenGL==3.1.6
pyparsing==3.0.9
python-dateutil==2.8.2
pytinyrenderer==0.0.14
PyYAML==6.0
rich==13.3.3
scipy==1.10.1
six==1.16.0
tensorboardX==2.6
tensorstore==0.1.35
toolz==0.12.0
trimesh==3.9.35
typing_extensions==4.5.0
Werkzeug==2.2.3
